refactor TerriaError + Result
